+++ src/main/java/com/ginjaninja/demoRestAPI/shiftAssignment/ShiftAssignmentService.java
@@ -1,9 +1,10 @@
-package com.ginjaninja.demoRestAPI.shift;
+package com.ginjaninja.demoRestAPI.shiftAssignment;
 
 import java.util.Date;
 
 import javax.transaction.Transactional;
 
+import org.joda.time.DateTime;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -30,14 +31,24 @@ public class ShiftAssignmentService {
     /**
      * Set default values for shiftAssignment and save
      * @param shiftAssignment    {@link ShiftAssignment}
-     * @return          {@link ShiftAssignment}
+     * @return       			 {@link ShiftAssignment}
      */
     public ShiftAssignment save(ShiftAssignment shiftAssignment){
-        if(shiftAssignment.getActiveInd() == null){
-            shiftAssignment.setActiveInd("Y");
-        }
-        this.updateActivityDtTm(shiftAssignment);
-        return shiftAssignmentDAO.save(shiftAssignment);
+    	if(this.validStartEnd(shiftAssignment.getStartDtTm(), shiftAssignment.getEndDtTm())){
+	        if(shiftAssignment.getActiveInd() == null){
+	            shiftAssignment.setActiveInd("Y");
+	        }
+	        if(shiftAssignment.getCreatedDtTm() == null){
+	        	shiftAssignment.setCreatedDtTm(new Date());
+	        }
+	        if(shiftAssignment.getCreatedDtTm() == null){
+	        	shiftAssignment.setCreatedDtTm(new Date());
+	        }
+	        this.updateActivityDtTm(shiftAssignment);
+	        return shiftAssignmentDAO.save(shiftAssignment);
+    	}else{
+    		return null;
+    	}
     }
     
     /**
@@ -47,16 +58,25 @@ public class ShiftAssignmentService {
      */
     public ShiftAssignment update(ShiftAssignment shiftAssignment) {
         this.fillShiftAssignment(shiftAssignment);
-        return shiftAssignmentDAO.update(shiftAssignment);
+        if(this.validStartEnd(shiftAssignment.getStartDtTm(), shiftAssignment.getEndDtTm())){
+        	return shiftAssignmentDAO.update(shiftAssignment);
+        }else{
+        	return null;
+        }
     }
     
     /**
      * Delete shiftAssignment with id
      * @param id        {@link Integer}
      */
-    public void delete(Integer id) {
+    public Boolean delete(Integer id) {
         ShiftAssignment shiftAssignment = shiftAssignmentDAO.get(id);
-        shiftAssignmentDAO.delete(shiftAssignment);
+        if(shiftAssignment == null){
+        	return false;
+        }else{
+        	shiftAssignmentDAO.delete(shiftAssignment);
+        	return true;
+        }	
     }
     
     /**
@@ -98,5 +118,21 @@ public class ShiftAssignmentService {
             shiftAssignment.setActivityDtTm(new Date());
         }
         return shiftAssignment;
+    }
+    
+    /**
+     * Validate start and end dates
+     * @param StartDtTm	Date
+     * @param endDtTm	Date
+     * @return			Boolean
+     */
+    private Boolean validStartEnd(Date StartDtTm, Date endDtTm){
+    	Boolean bool = true;
+    	DateTime start = new DateTime(StartDtTm);
+    	DateTime end = new DateTime(endDtTm);
+    	if(start.isAfter(end) || end.isBefore(start)){
+    		bool = false;
+    	}
+    	return bool;
     }
 }
